Revert "Clear code feed"
