+++ lib/app/common/common.dart
@@ -4,5 +4,4 @@ export 'theme/pngPath.dart';
 export 'theme/style.dart';
 export 'theme/svgPath.dart';
 export 'utils/text_utils.dart';
-export 'utils/list_game_utils.dart';
 export 'config/config.dart';
