+++ lib/app/pages/feeds.dart
@@ -1,6 +1,7 @@
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:carousel_slider/carousel_slider.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 import 'package:oneplay_flutter_gui/app/common/common.dart';
 import 'package:oneplay_flutter_gui/app/models/game_feed_model.dart';
@@ -78,6 +79,82 @@ class _FeedsState extends State<Feeds> {
                 ],
               ),
             )),
+    );
+    // return Container(
+    //   padding: const EdgeInsets.all(4),
+    //   alignment: Alignment.center,
+    //   child: Column(
+    //     children: [
+    //       Observer(builder: (_) {
+    //         return Text('Hi ${authService.user?.firstName}');
+    //       }),
+    //       const SizedBox(height: 24),
+    //       OutlinedButton(
+    //         onPressed: logginOut ? null : logout,
+    //         child: const Text('Logout'),
+    //       ),
+    //     ],
+    //   ),
+    // );
+  }
+
+  SizedBox listGames(GameFeedModel value) {
+    return SizedBox(
+      height: 127.59,
+      child: ListView(
+        scrollDirection: Axis.horizontal,
+        children: value.games
+            .map((e) => InkWell(
+                  onTap: (() => Modular.to.pushNamed('/game/${e.oneplayId}')),
+                  child: ColorFiltered(
+                    colorFilter: ColorFilter.mode(
+                        e.status == "live"
+                            ? Colors.transparent
+                            : Colors.grey.withOpacity(0.5),
+                        BlendMode.srcOver),
+                    child: Container(
+                      alignment: Alignment.center,
+                      margin: const EdgeInsets.only(right: 20),
+                      height: 127.59,
+                      child: ClipRRect(
+                        borderRadius: BorderRadius.circular(10),
+                        child: CachedNetworkImage(
+                          imageUrl: e.textBgImage.toString(),
+                          fit: BoxFit.fitHeight,
+                          height: 127.59,
+                          placeholder: (context, url) {
+                            return const Center(
+                              child: CircularProgressIndicator(),
+                            );
+                          },
+                          errorWidget: (context, url, error) {
+                            return Stack(
+                              children: [
+                                Image.asset(
+                                  defaultBg,
+                                  fit: BoxFit.fitHeight,
+                                ),
+                                Positioned(
+                                  top: 40,
+                                  left: 5,
+                                  child: Text(
+                                    e.title.toString(),
+                                    style: const TextStyle(
+                                        color: Colors.white,
+                                        fontFamily: mainFontFamily),
+                                    maxLines: 2,
+                                  ),
+                                ),
+                              ],
+                            );
+                          },
+                        ),
+                      ),
+                    ),
+                  ),
+                ))
+            .toList(),
+      ),
     );
   }
